version 0.2.2 with cross platform fixes
Avoid using Array.find or array.prototype.find. as IE 11 does not
support support the array.find() method. Alternatively a polyfill could
(see https://github.com/paulmillr/es6-shim) and
individual https://github.com/paulmillr/Array.prototype.find.

But found it easier to simply avoid this code.

IE 11 failed test when cookies are not stored.

The test that failed was:
''callback with bad signature for authorization response and jws-validator-decodeonly''

The symptom was that the cookie storing the encryption secret is not actually stored
and so deserialize() could not decrypt the session which then caused deserialize
to reset the session. The deserialise call is made when the localStorage event for
key 'anvil.connect' is triggered which is caused by calling serialize from the
Anvil.promise.callback method that the test is validating.

Now the reason the secret cooky was not stored is that the test mocks the time
so that the access token's time claims check out. However while this sets
an expire time of the cookie relative to the mocked time, it appears that IE
actually used the real time to let the cookie expire.

I have not checked whether this affected only IE. Another possibility is that
the timing of calling the promise handlers and the localStorage event
is different for the more modern browsers vs IE.
At any rate I don't think that deserialise should reset the session data.

Fixing this does change this test to succeed on all the tested browsers:

Safari 9.0.3 (Mac OS X 10.10.5): Executed 155 of 173 (skipped 18) SUCCESS (0.82 secs / 0.728 secs)
Firefox 43.0.0 (Mac OS X 10.10.0): Executed 155 of 173 (skipped 18) SUCCESS (0.904 secs / 0.781 secs)
Chrome 48.0.2564 (Mac OS X 10.10.5): Executed 155 of 173 (skipped 18) SUCCESS (0.748 secs / 0.674 secs)
IE 11.0.0 (Windows 10 0.0.0): Executed 155 of 173 (skipped 18) SUCCESS (0.776 secs / 0.787 secs)
Edge 13.10586.0 (Windows 10 0.0.0): Executed 155 of 173 (skipped 18) SUCCESS (1.894 secs / 1.801 secs)