Introduce not-authenticated event.
This is useful for angular to listen for logouts discovered via session management
so that it can invoke $rootScope.$apply().

Reverted a change when deserialize cannot restore a session to leave it alone.
Now instead it will set the session to an empty object and also update sessionState
from localStorage.
Furthermore it will emit the 'not-authenticated' event.

A subscriber may listen like so:

Anvil.on('not-authenticated, function(emptySession) {
   $rootScope.$apply()
})

Also reduced logging but left comments for checkSession to reduce noise.